JL8-20 refactor : actions파일 위치 이동, pages.tsx 컴포넌트 분리,중복 코드 제거
