+++ package.json
@@ -18,6 +18,7 @@
     "react": "^18",
     "react-bottom-sheet": "^1.1.1",
     "react-dom": "^18",
+    "react-icons": "^5.4.0",
     "react-player": "^2.16.0",
     "sweetalert2": "^11.15.10"
   },
